Disabling CS changes that break the test suite
In this specific occurrence, for whatever reason, tests around the
reflection capabilities of the library verify that there are two
parameters, with types declared in the docblock.

Native types seem to be completely ignored by the reflector, and thus
the test fails when the CS automation simplifies an import.

While it would be nice to get rid of this silly design, it's not
worth doing, and it is preferable to instead just suppress the
CS rule here, rather than introducing instability.